ref #2262

- Added support for GQA. Currently, the grouping factor (8 for 70Bv2, 1 for everything else) has to be passed from the command line: `-gqa 8`. With GGUF it will be read from the the model hparams
- The `ffn_dim_multiplier` needed to determine the correct value for `n_ff` is hardcoded to `1.3` when a 80 layer model with `GQA == 8` is loaded (i.e. this corresponds to 70Bv2). Otherwise, it defaults to `1.0`. Also needs to be read from the model hparams in the future
- CUDA support for GQA is provided by @JohannesGaessler 

```bash
# usage
python3 convert.py --outfile models/70B-v2/ggml-model-f16.bin --outtype f16 ../llama2/llama/llama-2-70b/
./quantize ./models/70B-v2/ggml-model-f16.bin ./models/70B-v2/ggml-model-q4_0.bin q4_0
./main -m ./models/70B-v2/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap --ignore-eos -n 64 -t 8 -gqa 8
```

Some notes while working on this:

I haven't done perplexity calcs yet, but I ran a few text generations using the small 7Bv2 model. My impression is that with this new model, the `Q4_0` and `Q5_0` quantizations do not work well. What I notice is that after the first sentence ends, the new sentence starts of in wild ways, often switching to German or some other language. This is something I have not observed with the original LLaMA 7B. It would also often start the second sentence without a capital letter, which again has never been the case before. 

The `QX_1`, `QX_K` and `Q8_0` quantizations do not seem to exhibit this behaviour (or if they do it is to a much smaller extend)

Note, the examples below are not related to the changes in this PR. The same behaviour is observed on `master` using the LLaMAv2 7B model.

Here are a few examples:

```bash
# command
./main -m ${model} -p "I believe the meaning of life is" --ignore-eos -n 32 -t 8

# Q4_0
 I believe the meaning of life is to find your purpose. sierpamid2015
I've always been a very spiritual person and I think the purpose of life is that we

 I believe the meaning of life is to be happy.☉
I am happy to tell you what I think about you.☆
Do I have a right to feel hurt when someone doesn'

 I believe the meaning of life is to find your passion and to have the courage to pursue it. nobody should stop you from living that dream if they want it for themselves...

 I believe the meaning of life is to help others and make people laugh. Einzeln ist das Leben nicht viel wert...
„In the past, it was a joke in Hollywood that if

# Q4_1
 I believe the meaning of life is to find your purpose.
Honorable Mentions: Astonishing, Incredible, Excellent, Impressive, and Fant

 I believe the meaning of life is to be a good parent, spouse and person.﻿
Kimberly is an Aussie and me a Kiwi but we’re married

 I believe the meaning of life is to find your passion and to have the courage to pursue it.
What was your first role as a kid? And who were you performing for?

 I believe the meaning of life is to learn and grow as a person. It's not about what you have in your bank account, how much money you make, or the size of your

# Q5_0
 I believe the meaning of life is to find your purpose. Hinweis auch, dass die Daten durch den Spieler aktuell gehalten werden müssen und somit regelmäßig gelösch

 I believe the meaning of life is to be a good man. Hinweis: Es ist keine göttliche oder universelle Antwort!
My 10-year old daughter recently asked me

 I believe the meaning of life is to find your passion and share it with others. nobody can tell you what's right for you in this life, if anything is wrong at all, it

 I believe the meaning of life is to learn and grow, but it's also about being happy. surely we don't need a million years to do that.
When you are young

# Q5_1
 I believe the meaning of life is to be a good example.
The following are just some short thoughts that have crossed my mind in relation to these scriptures:

 I believe the meaning of life is to find your passion and follow it.
The meaning of life is not happiness, but growth in a continuous journey

 I believe the meaning of life is to enjoy it. You’re only here for a short time, so do what you want!
The meaning of life in itself is not important. It

 I believe the meaning of life is to create and nurture an awareness that leads to self-love, authenticity and purpose.
My passion for this subject began with my

# Q4_K
 I believe the meaning of life is to create. You can make whatever you want in this world, whether it's a chair or a family. And if you think about it, everything is

 I believe the meaning of life is to enjoy it. Life has been a great teacher for me, and every day offers new lessons on how to live in peace and love.

 I believe the meaning of life is to find your passion and to love it. I think that we are here to learn, experience new things and grow so therefore if you aren’t doing what

 I believe the meaning of life is to be as happy, joyful and satisfied as you can throughout your time on earth. A life where, when you look back upon it at the end,

# Q8_0
 I believe the meaning of life is to be found in personal relationships. The only way for a relationship to grow and become meaningful, however, is that both parties are committed and devoted to it

 I believe the meaning of life is to find your purpose.
then fulfil it with a passion.
Finding my purpose was easy - but the journey that led me there, the less

 I believe the meaning of life is to learn and grow as a person. The meaning of life is to experience it, to see new things, meet new people, to discover new places, and

 I believe the meaning of life is to make it as beautiful and wonderful as possible.
This is a work in progress, but I’ll get there one day!
```

### Old description below

This works for 70B LLaMA-v2:

```bash
python3 convert.py --outfile models/70B-v2/ggml-model-f16.bin --outtype f16 ../llama2/llama/llama-2-70b/
./quantize ./models/70B-v2/ggml-model-f16.bin ./models/70B-v2/ggml-model-q4_0.bin q4_0
./main -m ./models/70B-v2/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap --ignore-eos -n 64 -t 8
```

<details>

```
$ ▶ make -j && gdb --args ./bin/main -m ../models/70B-v2/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap --ignore-eos -n 64 -t 8
[  2%] Generating build details from Git
[  6%] Built target ggml
Consolidate compiler generated dependencies of target llama
-- Found Git: /usr/bin/git (found version "2.34.1") 
[  8%] Built target ggml_static
[ 10%] Building CXX object CMakeFiles/llama.dir/llama.cpp.o
[ 10%] Built target BUILD_INFO
[ 12%] Linking CXX static library libllama.a
[ 12%] Built target llama
[ 14%] Linking CXX executable ../bin/test-quantize-fns
[ 19%] Linking CXX executable ../bin/test-grad0
[ 19%] Linking CXX executable ../bin/test-sampling
[ 21%] Built target common
[ 27%] Linking CXX executable ../bin/test-tokenizer-0
[ 27%] Linking CXX executable ../../bin/quantize
[ 27%] Linking CXX executable ../../bin/quantize-stats
[ 29%] Linking CXX executable ../bin/test-quantize-perf
[ 53%] Linking CXX executable ../../bin/server
[ 53%] Linking CXX executable ../../bin/embedding
[ 53%] Linking CXX executable ../../bin/benchmark
[ 53%] Linking CXX executable ../../bin/train-text-from-scratch
[ 53%] Linking CXX executable ../../bin/main
[ 53%] Linking CXX executable ../../bin/save-load-state
[ 53%] Linking CXX executable ../../bin/vdot
[ 53%] Linking CXX executable ../../bin/simple
[ 53%] Linking CXX executable ../../bin/baby-llama
[ 53%] Built target embdinput
[ 57%] Linking CXX executable ../../bin/perplexity
[ 57%] Linking CXX executable ../../bin/q8dot
[ 59%] Linking CXX executable ../../bin/embd-input-test
[ 61%] Built target test-grad0
[ 63%] Built target test-quantize-fns
[ 65%] Built target test-quantize-perf
[ 68%] Built target quantize
[ 70%] Built target vdot
[ 72%] Built target test-tokenizer-0
[ 76%] Built target test-sampling
[ 76%] Built target q8dot
[ 78%] Built target benchmark
[ 82%] Built target embd-input-test
[ 82%] Built target baby-llama
[ 85%] Built target main
[ 87%] Built target save-load-state
[ 89%] Built target embedding
[ 91%] Built target simple
[ 93%] Built target perplexity
[ 95%] Built target train-text-from-scratch
[ 97%] Built target quantize-stats
[100%] Built target server
GNU gdb (Ubuntu 12.1-0ubuntu1~22.04) 12.1
Copyright (C) 2022 Free Software Foundation, Inc.
License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>
This is free software: you are free to change and redistribute it.
There is NO WARRANTY, to the extent permitted by law.
Type "show copying" and "show warranty" for details.
This GDB was configured as "x86_64-linux-gnu".
Type "show configuration" for configuration details.
For bug reporting instructions, please see:
<https://www.gnu.org/software/gdb/bugs/>.
Find the GDB manual and other documentation resources online at:
    <http://www.gnu.org/software/gdb/documentation/>.

For help, type "help".
Type "apropos word" to search for commands related to "word"...
Reading symbols from ./bin/main...
(gdb) r
Starting program: /home/ggerganov/development/github/llama.cpp/build-rwdi/bin/main -m ../models/70B-v2/ggml-model-q4_0.bin -p I\ believe\ the\ meaning\ of\ life\ is --no-mmap --ignore-eos -n 64 -t 8
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
main: build = 852 (294f424)
main: seed  = 1689771309
llama.cpp: loading model from ../models/70B-v2/ggml-model-q4_0.bin
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32000
llama_model_load_internal: n_ctx      = 512
llama_model_load_internal: n_embd     = 8192
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 64
llama_model_load_internal: n_layer    = 80
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: freq_base  = 10000.0
llama_model_load_internal: freq_scale = 1
llama_model_load_internal: ftype      = 2 (mostly Q4_0)
llama_model_load_internal: n_ff       = 28672
llama_model_load_internal: model size = 70B
llama_model_load_internal: ggml ctx size = 37070.93 MB
llama_model_load_internal: mem required  = 40208.93 MB (+ 5120.00 MB per state)
llama_new_context_with_model: kv self size  = 1280.00 MB

system_info: n_threads = 8 / 32 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 0 | SSE3 = 1 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 64, n_keep = 0


 I believe the meaning of life is to give, not to get. I have found that the more we give our time, energy and love to others, the more fulfilled our lives become. Our relationships grow stronger and deeper.
I also believe that God has blessed each one of us with unique gifts and talents. It is up to
llama_print_timings:        load time = 14604.73 ms
llama_print_timings:      sample time =    25.78 ms /    64 runs   (    0.40 ms per token,  2483.03 tokens per second)
llama_print_timings: prompt eval time =  5393.69 ms /     8 tokens (  674.21 ms per token,     1.48 tokens per second)
llama_print_timings:        eval time = 57526.02 ms /    63 runs   (  913.11 ms per token,     1.10 tokens per second)
llama_print_timings:       total time = 62954.88 ms
[Inferior 1 (process 1698964) exited normally]
(gdb) q
```
</details>

Currently, only CPU.
This is very quick and dirty - just to see what changes are necessary.

Good news is the `convert.py` script does not require changes.
To add support for GPU and BLAS, fix the following TODOs:

https://github.com/ggerganov/llama.cpp/blob/2d2bb6b0af7df316c297ec0fbc9c8af3566c23f1/ggml.c#L10729-L10748

I.e., implement the `mul_mat` broadcast logic from here:

https://github.com/ggerganov/llama.cpp/blob/2d2bb6b0af7df316c297ec0fbc9c8af3566c23f1/ggml.c#L10832-L10850

Looking for contributions to make this cleaner. If not, I will implement it some time in the next days.