@SlyEcho 2023-07-19
So it seems another model parameter is needed, group size? Which is 1 for the older LLaMA models and 8 for the new ones?