@ggerganov 2023-07-19
Maybe try to deduce it for now, to avoid file format mess. We'll add it in GGUF