@SlyEcho 2023-07-19
Do I understand correctly that this mechanism will reduce the KV cache size as well?