@JohannesGaessler 2023-07-19
In terms of CUDA support: for me the release has come at a rather inopportune time. I'll maybe have some time to look into it on the weekend but I don't want to make any promises. Other people are welcome to work on it in the meantime (but if you do please notify me to avoid duplicate implementations).