@byildiz 2023-07-19
> Do I understand correctly that this mechanism will reduce the KV cache size as well?

Yes, the mechanism (GQA) is solely for reducing KV cache size by compromising some amount of accuracy.