@gabrielcustodio 2023-07-19
How to run ./quantize ?