@wizzard0 2023-07-19
@gabrielcustodio 
```
git clone https://huggingface.co/meta-llama/Llama-2-70b

python convert.py Llama-2-70b

./quantize Llama-2-70b/ggml-model-f32.bin Llama-2-70b/ggml-model-q5_1.bin q5_1

./main -m Llama-2-70b/ggml-model-q5_1.bin --interactive-first
```