@nrbontha 2023-07-20
@wizzard0 

Thank you for clearing up the steps - however, I'm seeing this error:

`$ python3 convert.py Llama-2-70b
Loading model file Llama-2-70b/consolidated.00.pth
Traceback (most recent call last):
  File "/Users/nrb/Projects/llama.cpp/convert.py", line 1264, in <module>
    main()
  File "/Users/nrb/Projects/llama.cpp/convert.py", line 1244, in main
    model_plus = load_some_model(args.model)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/nrb/Projects/llama.cpp/convert.py", line 1165, in load_some_model
    models_plus.append(lazy_load_file(path))
                       ^^^^^^^^^^^^^^^^^^^^
  File "/Users/nrb/Projects/llama.cpp/convert.py", line 963, in lazy_load_file
    raise ValueError(f"unknown format: {path}")
ValueError: unknown format: Llama-2-70b/consolidated.00.pth`

Looking at the `convert.py` script, it seems that `.pth` files are not supported - but I'm new here and may be missing a step. 

Anything I can try to overcome this?