@SlyEcho 2023-07-20
@nrbontha, .pth files are supported from the start of llama.cpp so I think the problem you have is that the files you have is corrupted.