@SlyEcho 2023-07-20
Meta's script also downloads MD5 checksums, but just check if the file is not 0 bytes, I had that problem when downloading.