@schappim 2023-07-20
> 

Thanks for trying to help @SlyEcho, but I'm afraid that isn't the issue (at least for me): 

```
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:12 consolidated.00.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:12 consolidated.01.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:12 consolidated.02.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:12 consolidated.03.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:12 consolidated.04.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:14 consolidated.06.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:15 consolidated.05.pth
-rw-r--r--@  1 admin  staff  17246706245 14 Jul 09:15 consolidated.07.pth
```