@l0d0v1c 2023-07-20
I got this error as an output from main in apple silicon
`llama_model_load_internal: ggml ctx size = 37070,93 MB
error loading model: llama.cpp: tensor 'layers.0.attention.wk.weight' has wrong shape; expected  8192 x  8192, got  8192 x  1024
llama_load_model_from_file: failed to load model
llama_init_from_gpt_params: error: failed to load model 'llama-2-70B-chat.bin'
main: error: unable to load model`