@klosax 2023-07-20
> I got this error as an output from main on apple silicon
> `llama_model_load_internal: ggml ctx size = 37070,93 MB error loading model: llama.cpp: tensor 'layers.0.attention.wk.weight' has wrong shape; expected 8192 x 8192, got 8192 x 1024 `

@l0d0v1c  Are you compiling with the changes applied that this PR propose?
https://github.com/ggerganov/llama.cpp/pull/2276/files