@schappim 2023-07-20
> ```shell
> -m ./models/70B-v2/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap --ignore-eos -n 64 -t 8
> ```

🎉 Success! I got it working by using cmake instead of make , adn then using the outputs in ./build/bin/ . 


```
git checkout llama-v2-70b;
python3 convert.py --outfile models/70B-v2/ggml-model-f16.bin --outtype f16 ../llama2/llama/llama-2-70b/ ;
./build/bin/quantize  ./models/70B-v2/ggml-model-f16.bin ./models/70B-v2/ggml-model-q4_0.bin q4_0;
./build/bin/main -m ./models/70B-v2/ggml-model-q4_0.bin -n 256 --repeat_penalty 1.0 --color -i -r "User:" -f ./prompts/chat-with-bob.txt
```
Now remember this is running on the CPU, so be prepared for some chugging!

<img width="1053" alt="image" src="https://github.com/ggerganov/llama.cpp/assets/40018/f4501aa4-9c2c-4176-8c0d-28b4ace8ac30">

Y'know what I call this? I call this a good start!