@klosax 2023-07-20
It looks like changes are needed in convert.py to work for 70b HF files as the [`find_n_mult()`](https://github.com/ggerganov/llama.cpp/blob/294f424554c1599784ac9962462fc39ace92d8a5/convert.py#L134C1-L134C48) function wont work with this model. 
See issue https://github.com/ggerganov/llama.cpp/issues/2286

This function convert the `feedforward_length` parameter back to the original pth `n_mult` for saving. When the file is load it is converted back again... Luckily hacks like this wont be needed with the new GGUF file format.