@SlyEcho 2023-07-20
> I got it working by using cmake instead of make

Remember to run `make clean` first or use the `-B` flag.