@klosax 2023-07-20
Getting garbage output from q4_0 converted from safetensor model https://huggingface.co/TheBloke/Llama-2-70B-fp16, but works fine with q4_0 converted from (original?) .pth model https://huggingface.co/anonymous4chan/llama-2-70b-original .

The safetensor model files were downloaded ok (checked sha256sums) so guessing they are corrupt or somehow incompatible with the convert.py script..