@klosax 2023-07-21
Tried to convert another 70b safetensor model https://huggingface.co/NousResearch/Llama-2-70b-hf and quantized it to q4_0 and it also outputs garbage. An old llama version 1 7b safetensor model works fine.

Is the current convert.py incompatible with the llama2-70b safetensor files?
Anyone else having this problem?