@auxon 2023-07-21
> 

@schappim What hardware are you using?  I tried the 13b-chat model on my M1 Macbook Air, but it is incredibly slow.