@JohannesGaessler 2023-07-22
The CUDA implementation already works as-is if you only offload the repeating and non-repeating layers but not the KV cache.

<details>
```
   ~/Pr/llama.cpp    llama-v2-70b wip *4 ?12    ./main --model models/opt/llama_2-70b-ggml-q4_k_m --seed 1338 --ignore-eos --n-predict 128 --ctx-size 2048 --batch-size 512 --threads 16 --prompt "Llamas are animals that " -ngl 81 | tee chat.txt
main: build = 853 (2d2bb6b)
main: seed  = 1338
ggml_init_cublas: found 3 CUDA devices:
  Device 0: Tesla P40, compute capability 6.1
  Device 1: Tesla P40, compute capability 6.1
  Device 2: Tesla P40, compute capability 6.1
llama.cpp: loading model from models/opt/llama_2-70b-ggml-q4_k_m
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32000
llama_model_load_internal: n_ctx      = 2048
llama_model_load_internal: n_embd     = 8192
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 64
llama_model_load_internal: n_layer    = 80
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: freq_base  = 10000,0
llama_model_load_internal: freq_scale = 1
llama_model_load_internal: ftype      = 15 (mostly Q4_K - Medium)
llama_model_load_internal: n_ff       = 28672
llama_model_load_internal: model size = 70B
llama_model_load_internal: ggml ctx size =    0,19 MB
llama_model_load_internal: using CUDA for GPU acceleration
ggml_cuda_set_main_device: using device 0 (Tesla P40) as main device
llama_model_load_internal: mem required  = 3476,82 MB (+ 5120,00 MB per state)
llama_model_load_internal: allocating batch_size x (1536 kB + n_ctx x 516 B) = 1284 MB VRAM for the scratch buffer
llama_model_load_internal: offloading 80 repeating layers to GPU
llama_model_load_internal: offloading non-repeating layers to GPU
llama_model_load_internal: offloaded 81/83 layers to GPU
llama_model_load_internal: total VRAM used: 40607 MB
llama_new_context_with_model: kv self size  = 5120,00 MB

system_info: n_threads = 16 / 32 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1,100000, presence_penalty = 0,000000, frequency_penalty = 0,000000, top_k = 40, tfs_z = 1,000000, top_p = 0,950000, typical_p = 1,000000, temp = 0,800000, mirostat = 0, mirostat_lr = 0,100000, mirostat_ent = 5,000000
generate: n_ctx = 2048, n_batch = 512, n_predict = 128, n_keep = 0


 Llamas are animals that 50% of the world’s population is aware of.
While they are native to South America, they have been seen all over the globe in recent decades as people become more interested in this species of animal.
However, when most people think about llamas and South America, the first thing that comes up is Peru.
But how much do you know about the llama? Or even just about where they come from and how to take care of them properly? In this article, we will give a brief introduction on everything there is to know about llamas in Peru.
Where Does The Ll
llama_print_timings:        load time = 11149,64 ms
llama_print_timings:      sample time =    77,85 ms /   128 runs   (    0,61 ms per token,  1644,19 tokens per second)
llama_print_timings: prompt eval time =  1139,59 ms /     8 tokens (  142,45 ms per token,     7,02 tokens per second)
llama_print_timings:        eval time = 19265,68 ms /   127 runs   (  151,70 ms per token,     6,59 tokens per second)
llama_print_timings:       total time = 20522,62 ms
```
</details>